move format functions out of psc controller to format util
