+++ src/controllers/tasks/people.with.significant.control.controller.ts
@@ -19,7 +19,7 @@ import { Session } from "@companieshouse/node-session-handler";
 import { getPscs } from "../../services/psc.service";
 import { createAndLogError, logger } from "../../utils/logger";
 import { toReadableFormat } from "../../utils/date";
-import { formatTitleCase, formatAddressForDisplay } from "../../utils/format";
+import { formatServiceAddress, formatPSCForDisplay, formatUraAddress } from "../../utils/format";
 import { sendUpdate } from "../../utils/update.confirmation.statement.submission";
 
 export const get = async (req: Request, res: Response, next: NextFunction) => {
@@ -125,61 +125,3 @@ const getPscStatementUrl = (req: Request, isPscFound: boolean) => {
   return urlUtils.setQueryParam(path, URL_QUERY_PARAM.IS_PSC, isPscFound.toString());
 };
 
-const formatPSCForDisplay = (psc: PersonOfSignificantControl): PersonOfSignificantControl => {
-  const clonedPsc: PersonOfSignificantControl = JSON.parse(JSON.stringify(psc));
-  if (psc.nameElements) {
-    clonedPsc.nameElements = {
-      forename: formatTitleCase(psc.nameElements?.forename),
-      otherForenames: psc.nameElements?.otherForenames,
-      surname: psc.nameElements?.surname,
-      middleName: psc.nameElements?.middleName,
-      title: psc.nameElements?.title
-    };
-  }
-
-  if (psc.address) {
-    clonedPsc.address = {
-      addressLine1: formatTitleCase(psc.address.addressLine1),
-      addressLine2: formatTitleCase(psc.address.addressLine2),
-      careOf: formatTitleCase(psc.address.careOf),
-      country: formatTitleCase(psc.address.country),
-      locality: formatTitleCase(psc.address.locality),
-      poBox: formatTitleCase(psc.address.poBox),
-      postalCode: psc.address.postalCode,
-      premises: formatTitleCase(psc.address.premises),
-      region: formatTitleCase(psc.address.region)
-    };
-  }
-
-  clonedPsc.serviceAddressLine1 = formatTitleCase(psc.serviceAddressLine1);
-  clonedPsc.serviceAddressPostTown = formatTitleCase(psc.serviceAddressPostTown);
-  clonedPsc.serviceAddressCareOf = formatTitleCase(psc.serviceAddressCareOf);
-  clonedPsc.serviceAddressCountryName = formatTitleCase(psc.serviceAddressCountryName);
-  clonedPsc.serviceAddressPoBox = formatTitleCase(psc.serviceAddressPoBox);
-  clonedPsc.serviceAddressArea = formatTitleCase(psc.serviceAddressArea);
-  clonedPsc.serviceAddressRegion = formatTitleCase(psc.serviceAddressRegion);
-
-  return clonedPsc;
-};
-
-const formatServiceAddress = (formattedPsc: PersonOfSignificantControl): string => {
-  const addressClone: Address = {
-    addressLine1: formattedPsc.serviceAddressLine1,
-    careOf: formattedPsc.serviceAddressCareOf,
-    country: formattedPsc.serviceAddressCountryName,
-    locality: formattedPsc.serviceAddressArea,
-    poBox: formattedPsc.serviceAddressPoBox,
-    postalCode: formattedPsc.serviceAddressPostCode,
-    premises: formattedPsc.serviceAddressPostTown,
-    region: formattedPsc.serviceAddressRegion
-  };
-  return formatAddressForDisplay(addressClone);
-};
-
-const formatUraAddress = (formattedPsc: PersonOfSignificantControl): string => {
-  let ura = "";
-  if (formattedPsc.address) {
-    ura = formatAddressForDisplay(formattedPsc.address);
-  }
-  return ura;
-};
