+++ src/utils/format.ts
@@ -1,4 +1,4 @@
-import { ActiveDirectorDetails, Address } from "@companieshouse/api-sdk-node/dist/services/confirmation-statement";
+import { ActiveDirectorDetails, Address, PersonOfSignificantControl } from "@companieshouse/api-sdk-node/dist/services/confirmation-statement";
 
 export const formatTitleCase = (str: string|undefined): string =>  {
   if (!str) {
@@ -48,4 +48,63 @@ export const formatAddress = (address: Address): Address => {
     region: formatTitleCase(addressClone.region),
     postalCode: addressClone.postalCode?.toUpperCase()
   };
+};
+
+export const formatPSCForDisplay = (psc: PersonOfSignificantControl): PersonOfSignificantControl => {
+  const clonedPsc: PersonOfSignificantControl = JSON.parse(JSON.stringify(psc));
+  if (psc.nameElements) {
+    clonedPsc.nameElements = {
+      forename: formatTitleCase(psc.nameElements?.forename),
+      otherForenames: psc.nameElements?.otherForenames,
+      surname: psc.nameElements?.surname,
+      middleName: psc.nameElements?.middleName,
+      title: psc.nameElements?.title
+    };
+  }
+
+  if (psc.address) {
+    clonedPsc.address = {
+      addressLine1: formatTitleCase(psc.address.addressLine1),
+      addressLine2: formatTitleCase(psc.address.addressLine2),
+      careOf: formatTitleCase(psc.address.careOf),
+      country: formatTitleCase(psc.address.country),
+      locality: formatTitleCase(psc.address.locality),
+      poBox: formatTitleCase(psc.address.poBox),
+      postalCode: psc.address.postalCode,
+      premises: formatTitleCase(psc.address.premises),
+      region: formatTitleCase(psc.address.region)
+    };
+  }
+
+  clonedPsc.serviceAddressLine1 = formatTitleCase(psc.serviceAddressLine1);
+  clonedPsc.serviceAddressPostTown = formatTitleCase(psc.serviceAddressPostTown);
+  clonedPsc.serviceAddressCareOf = formatTitleCase(psc.serviceAddressCareOf);
+  clonedPsc.serviceAddressCountryName = formatTitleCase(psc.serviceAddressCountryName);
+  clonedPsc.serviceAddressPoBox = formatTitleCase(psc.serviceAddressPoBox);
+  clonedPsc.serviceAddressArea = formatTitleCase(psc.serviceAddressArea);
+  clonedPsc.serviceAddressRegion = formatTitleCase(psc.serviceAddressRegion);
+
+  return clonedPsc;
+};
+
+export const formatServiceAddress = (formattedPsc: PersonOfSignificantControl): string => {
+  const addressClone: Address = {
+    addressLine1: formattedPsc.serviceAddressLine1,
+    careOf: formattedPsc.serviceAddressCareOf,
+    country: formattedPsc.serviceAddressCountryName,
+    locality: formattedPsc.serviceAddressArea,
+    poBox: formattedPsc.serviceAddressPoBox,
+    postalCode: formattedPsc.serviceAddressPostCode,
+    premises: formattedPsc.serviceAddressPostTown,
+    region: formattedPsc.serviceAddressRegion
+  };
+  return formatAddressForDisplay(addressClone);
+};
+
+export const formatUraAddress = (formattedPsc: PersonOfSignificantControl): string => {
+  let ura = "";
+  if (formattedPsc.address) {
+    ura = formatAddressForDisplay(formattedPsc.address);
+  }
+  return ura;
 };
